Move the mongo collection definition to the records
During ingestion the mongo collection a record should be inserted into
or updated in is now presented from the record itself rather than just
being passed to the ingestion class on initialisation. This is helpful
as it allows potentially multiple collections to be modified by one
stream of records. The stats reporting has also been updated to reflect
this change.