Significant refactor of indexing code
This refactor is pretty large as it changes the fundamental concepts on which
the indexing process is built. Now, a series of objects dictate which index a
given data item are indexed into using filters, rather than one converter for
all. This has advantages in terms of subclassing and extending the functionality
where there isn't a one-to-one mapping from mongo collection to elasticsearch
index.