Refactor the ingestion time/start time to make it more understandable
It wasn't very clear which parameters would be used for what. There was
an implication that the start parameter passed to the ingester should be
the same as the ingester_time parameter that the converter gets, now it's
clear. This change also resolves another potential problem whereby we
calculate the duration of the ingestion incorrectly as we use the start
parameter passed to us. We keep this in-house now.