Refactor indexing, again
These change step back to a one-by-one model where
each mongo_doc is handled in isolation. This has mostly
been spurred on by work in the data-importer2 repo which
is currently the only user of this library. The changes
have resulted in an easier to follow structure and code
that should be more efficient too (even if it's only by
a very small amount).