Correctly add the commands individually to the op buffer
Before this change all the commands for a mongo doc were being added
at once which meant that when we attempted to create the data string
for elasticsearch we failed as we only unpacked one level of iterable
not two. This fixes that. It also ensures that the correct number of
commands sent in a bulk request to elasticsearch is abided by.