Allow keys to be ignored during diffing
This allows the records to specify if any keys should be ignored during
the diffing stage. An example usage of this would be if a record subclass
wishes to include a modification time in the data dict. Say, for example,
the modification time is altered but no parts of the record that actually
get stored in mongo are changed. In this scenario we probably don't want
a new version of the record to be created and therefore the modification
key needs to be ignored.