Switch all elasticsearch usage to the elasticsearch official library
It makes sense to use this over making the requests ourselves using
requests for 2 reasons: firstly, we won't be tied to one server in the cluser
as we're not just hitting one address and hoping it's there, the library
round robins the known hosts and discovers them periodically too.
Secondly, it makes it easier to use the query-dsl library elasticsearch
provide when we come to implementing search.