Bulk write to Elasticsearch in a thread
This commit alters how we push the documents to index to elasticsearch by
sending them in a thread on the end of a queue. Doing this works from a
performance point of view as the GIL is given up when the socket writes
the data to the elasticsearch server allowing other threads to continue,
i.e. the main thread that is processing the mongo documents.