Allow data and meta to be missing
If the searches doesn't specify at least a meta or data field in the _source of the
search then they may be missing. We could combat this by adding them in automatically
during the pre_search phase, however they aren't currently actually used by the eevee
core and therefore it seems a bit draconian to do that.
Also, adding them in could be a bit fiddly considering there are 3 different ways of
specifying the _source in a search and they can all use wildcards.