Flip the keyword and text definitions
It makes more sense for the keyword to be stored at the name of field and then
text type to be stored in a subfield. This allows us to do the primary jobs of
filtering and sorting without appending .keyword to everything. If fuzzy matching
is required then this is going to require a different setup any so adding .text
as well makes sense. This change also makes it easier to handle any subclasses
defining different types for fields as otherwise code using the fields would have
to be aware that field `a` was a date whereas field `b` was a text type and therefore
it should append .keyword to field `b` but not field `a` which is messy.