Remove stats collections
The indexer and ingester classes now just return the stats dicts
and don't insert them into the mongo database. This change has been
made so that users of the library can decide what to do with the
statistics generated (particularly given that the core eevee lib
wasn't using them for anything).