Report the targets in the indexing stats
Adds a "targets" field into the stats reports so that you don't
have to use the operations dict to figure out which indexes were
targetted in the index run. Note that this commit also changes
"source" to "sources" to reflect that it is a list (even if it
only contains one value).