Properly define the index and finish signals
In the non-multiprocess indexer class we could implement an
index signal which gets the mongo doc and list of commands
each time a record is handled, however in the multiprocess
version we can't do this as we'd have to pass the mongo doc
and commands back through to the main process which would
then trigger the signal. This will be horrible from a
performance angle which is what we're shooting for with the
multiprocessing variant. Therefore, only the counts are sent
and no more information. This ensures that the index signal
receives the same parameters regardless of which backend
indexer class is in use.