Fix multiprocessing indexing
Having blinker signals breaks the multiprocess indexer because they
hold un-picklable weak references. To solve this issue, the main
Indexer object is never sent through to the worker process in any
form. To achieve this the index_process function is moved out of
the class and initialised with the couple of parameters it needs
from the instance. This solves the issue. There are also a couple
of renames in of signal kwargs which were incorrect too.