Add a number type to the standard multifield
This allows field values to be indexed as numerical values which can then be
accessed and searched on using numerical search types (like range) by using
<field_name>.number.

This overcomes the "all values are stored as strings" shortcoming of the
system without causing issues with changing field data types between version.
For example, imagine a field called "Category" which intially has a vocabulary
of the numbers 1-5. If we indexed based on the values we'd create an integer
colour for this field. But what happens if in the next version of the data
this field's vocabulary has been extended to include the letters a-c. We now
have a problem as these will be rejected by the field's mapping. The only
way to solve this would be to do delete the index and completely reindex.
Using strings solves this problem and then using multifields improves the
functionality further.

A couple of other notes:

    - the double type has been used to ensure we can accept all numbers.
      We could probably use float, but this is more future proof. Additionally,
      elasticsearch only uses as much storage space as necessary for the values
      in the fields.

    - we could add other multifield types, such as date and boolean but these are
      less useful than number in the initial work (for example, it is easy to filter
      and search for the string value to emulate a boolean search). Also, boolean
      currently doesn't accept the "ignore_malformed" value which is problematic.