Expand the diffing capabilities and add a faster differ
If the data going through the diffing process is a shallow dict
(i.e. one without any nested dicts within it) then storing a diff
can be faster if we produce a diff in the form of another dict
which then when applied using dict.update() produces the patched
dict. This is faster because dict.update is part of the core lib
and therefore will be implemented in c.