Don't provide the id when indexing a record
This is a speed improvement as elasticsearch will autogenerate the id
and not have to check if the id we've provided is unique. We don't need
to provide the id anymore as we're not using it for anything - when we
reindex records we delete all the existing versions first using the
record id not the elasticsearch document ids.