Reorganise the searcher's version related helper functions
There were a few functions providing various functionality such as
getting all the versions available for an index, or getting all the
versions available for a specific record. This commit just tidies these
up so that their names are sensible and ensures we have helpers for all
the commonly needed functions.