Keep track of the versions we see during indexing
This is needed because we should update the latest version in the status
index with the actual latest version seen not the version passed through
to the constructor in case we see versions that are greater than the one
passed in.

Also, it means we can store it in the stats collection and that's nice
too.