Update indexer to account for "leftovers" bug
Under ideal circumstances, the same number of index documents, or more,
will be produced each time a record is indexed. In reality, things get
messed up and out of sync and we need to handle that better in eevee.
Therefore, we now check what currently exists in elasticsearch for a
given record and then update these as appropriate. This can involve:

  - leaving an index document alone because it is the same as a document
    we have generated during indexing
  - replacing an index document because it has changed
  - deleting a document because it is no longer correct and will not be
    replaced

This also has the added advantage of reducing our writing to
elasticsearch to a minimum as we only delete/replace as needed.