Add an always_replace option to override the write-saving noops
By default we avoid writing to elasticsearch to replace an index
document with the exact same document, however sometimes this is desired
(for example if there has been a mapping change and the index needs to
be reindexed). To facilitate this through the indexer in eevee (though
there are other ways to make a mapping change realised) we allow the
overriding of the comparison functionality.