Switch to using streaming_bulk instead of parallel_bulk
Few reasons for this, firstly it makes everything a bit simpler as we
eliminate all threading from the eevee lib. Secondly, it makes handling
exceptions a bit easier as we don't have to override the parallel_bulk
implementation from the elasticsearch helpers lib and modify how it
works. Thirdly, and finally, in testing it seems to improve the speed,
at least under some circumstances. We may wish to revisit this and add
the option to use threads/processes or not but equally with everything
now implemented single threaded, users can wrap their own
multi-threading/multi-processing around eevee if they really want to.