fix: fix tuple <-> dict value change diffing
If a value during diffing is identified as a tuple or a dict and it changes to the other container type (i.e. tuple becomes dict, dict becomes tuple) then we need to just log the change, not queue the values for diffing on the queue. This logic is now fixed and there is a test for it now too.