fix: change the MetaField enum to not use full paths as values
The values were the full paths which was confusing and also meant that a lot of the usages were wrong. Didn't break the tests though because we were referncing the enum all over the place, which is kind of the point of having it, so that's nice. To get the full path of the MetaField you're intested in, just call path() on it.