fix: ensure root field usage is consistent in generated field paths
In the fields module, there are a set of helpers which create field paths. These included the root in the path but then the resulting paths were being used in two different places, one where the root path was needed (the index template) and one where it was not (the parsing, specifically geo). The fields module helpers now have no root paths included in their returns and where they were needed in the index template they are just manually added.