feat: change the date field to use epoch_millis
Doesn't really matter what format we use here because we are completely controlling the value that is passed into this type of field, but using epoch_millis feels sensible as it's the underlying storage format Elasticsearch uses for date type values anyway so it doesn't have to do any additional conversions, and avoids any confusion with how timezones etc are handled. It's just clean and unambiguous. And it's a date type so you can search it with whatever you like anyway so this is purely a backend representation no one has to know anything about to use.