feat: change the prepare function to produce other simple types beyond str
BREAKING CHANGE: change the prepare function to produce other simple types beyond str

The prepare function now produces ints, floats, and bools as well as strs and Nones. This means we don't have to do any conversion on this data before putting it into mongodb which reduces the potential for errors when converting floats to strings and then back to floats (to string on the way into mongo, back to float when indexing). This also means we don't need a True -> "true" -> True cycle either. Finally, this also means that type changes trigger a diff change. This isn't ideal because it means we'll trigger a new version of the record even though the elasticsearch record is unlikely to actually be different. But it means we get some type information stored in mongo which we may want to use at some point and the only downside is that we produce more docs in elasticsearch than we technically need.