feat: add parsing configs for versioned control
Parsing configs allow users to control how their data is parsed and maintain the correct index parsing approach at any version. This is necessary because the data in elasticsearch is meant to be rebuildable, meaning we need to be able to reindex a record in the same way even if the configuration for how that record is indexed has changed (e.g. using different geo fields, using different date formats).