feat: change how uncommitted record data is handled
BREAKING CHANGE: change how uncommitted record data is handled

Instead of committing using a new version and then committing by moving that version into a status object, now the new data is added using a version of None and then when the changeset is complete it is committed by setting a new version on all of those records. This is much cleaner and simpler. This commit also includes small changes to accommodate this change in the syncing and some rollback features which are as yet mostly untested.