feat: add the meta.geo field back in, populated with all other record geo values
meta.geo existed in the old splitgill and is useful so makes a return. It is populated with a GeoJSON GeometryCollection containing all the GeoJSON values included in the reocrd's top-level geo field. Having this meta.geo field allows easy access to the record's geo data regardless of where which fields it actually came from.