fix: switch the diffing from using tuples to lists
BREAKING CHANGE: switch the diffing from using tuples to lists

MongoDB + pymongo stores tuples as lists and there seems to be no way to get those tuples back out as lists (pymongo provides custom type handlers, but you can't override the ones for builtin classes (such as list). With this in mind, it makes sense to alter our diffing code so that it doesn't rely on tuples as there are downstream consequences when things are tuples before they go into MongoDB but not when they come out, especially to patching.

Additionally, a refactor has been done to make following how dicts and lists are compared and DiffOps generated easier.