fix: allow the profiles index to use many more fields
This is not a good long term solution, but it will serve for now. Better solutions would be to either make the fields a list of dicts (probably the better solution) or set the mapping up so that it doesn't get indexed at all so loads of dynamic fields aren't created.