feat: refactor field definitions and add additional parsing options
This is an unfortunately large commit, sozz.It includes a move away from the enum field structure to strings with helper functions (though a single enum is maintained), along with the changes to everything using those fields as required. The search module interface is refactored because of this too. Additionally, because I am disorganised and lacking time, this commit includes the addition of 2 new parsing options, the keyword length and the float format.