feat: add a modified field option that can be ignored during mongo diff adds
When adding a new version of a record, this optional parameter allows for a modified field in the record's data to be ignored and not trigger the creation of a new version if it is the only change in the record data. This covers off the case where a record is touched and modified date in the record updated, but nothing else changes. There's no point in making a diff in this circumstance and this parameter allows us to deal with it. By default no modified field is set because it is unknown what the field name would be in all Splitgill use cases and not everyone might want this feature.