fix: stop creating new versions of records that are the same but have lists
Eurgh, remnant of a previous change where I hadn't propagated all of the change through the code. Other commits to follow tidying this up. But, what does this do? Well. There was some code which converted lists in record data that had come out of mongo to tuples, but we stopped using tuples as the primary container type and switched to lists, but I forgot to remove the code that did this list -> tuple switch. Thus if you tried adding the same data such as {"x": [1,2,3]} back to back into mongo you'd get two versions as the second time the data was ingested it would be compared against {"x": (1,2,3)}.