feat: introduce our own elasticsearch bulk op implementation
The elasticsearch lib's bulk helper implementation is pretty slow from testing. This isn't their fault, it's written to be flexible, but we know exactly what kind of operation we're dealing with and can therefore write a more efficient code which only deals with the use case we have. Additionally, the lib's helper code uses threads whereas our reimplementation uses asyncio. In reality this probably doesn't make that much difference in terms of how the GIL is managed and such, but again, writing it ourselves gives us more control over how to scale and manage the data flows with our specific use case in mind.