feat: rework our index doc structure to better accommodate geo + many knock on effects
BREAKING CHANGE: rework our index doc structure to better accommodate geo + many knock on effects

Many things in this commit all stemming from the route problem that on the free version of elasticsearch, you can't use geo tile aggregations on geo_shape values. This means we need to index shapes and points so that we can search on shapes or points, but aggregate on points, yay! To implement this, quite a lot has been reworked as well, mainly based around changing the structure of the index document to clean it up a bit, which obviously had even more knock on effects. Additional big change unfortunately buried in this commit is removal of profiles and the addition of a get_fields method on the database class. This is because a) profiles were a bit meh, b) I realised my approach didn't work downstream (the key realisation was data types != parsed types and trying to conflate them was causing major issues), and c) the new get_fields method is fast enough that you can run it dynamically instead of profiles which were an upfront, once per version job. This is nice because it means you can profile the database you're searching at any version with any filters. To generate the get_fields response, we just have to do 2 terms aggregation calls which use the indexes directly and are therefore really quick for es to execute. Yay!