fix: also inspect the document's next field when getting the current elasticsearch version
This change also updates the tests to be better as they were mocking the documents and actually not testing much. Using next is necessary because if a version in mongodb only contains deletes then this will only be noticed in the next field not version because there won't be a document in latest with a version.