fix: switch caret in field names to underscore
BREAKING CHANGE: switch caret in field names to underscore

Carets are a special character in certain kinds of elasticsearch operations so we need to use something else. This change switches them for underscores and introduces splitgill having underscore as a namespace indicator meaning fields can't start with underscore, except one's splitgill is using. This includes _id as the record id in the data which this commit introduces as well.

Closes: #20