feat: give arc documents an auto id and proactively delete on resync
BREAKING CHANGE: give arc documents an auto id and proactively delete on resync

Previously, documents that went into the arc indices got an ID which included the record ID and the version. This meant old documents could be found by version if needed and because all old versions of a record went into the same arc index, they could be replaced easily during a resync without issue. Now that old versions of a record could end up in any arc index, we need to proactively delete them when a resync is requsted. The latest index is fine and can stay as is, but the arc indices must be deleted to avoid duplicates. Because of this change, we also now don't need the arc documents to have IDs which means we can let elasticsearch set them whichs is faster on ingest :)