fix: correct index op generation logic
The specifics of this issue relate to a specific circumstance where between two versions of a record, an options change is made which doesn't impact the representation of the data at the earlier version. Due to a logic error in the index op generation function, this resulted in missing versions (specifically the newer version in this scenario) due to the ops being incorrectly generated. This has now been fixed in this commit by rewriting the index op generation function. Two specific parts of the rewrite are worth highlighting. Firsly, an effort has been made to make the code easier to read and reason about which is particularly useful given it's a fairly complex area of logic. Secondly, the original logic was very iterator heavy and traversed the record's versions backwards from newest to oldest. This was origianlly implemented as it matches the storage mechanism in mongo for the versions and as it was seen as being the most efficient way to do it. Not only did it save memory it also allowed us to easily stop at the point where the versions are already synced to Elasticsearch (the after parameter) and not generate more data than we need. In the new version, all versions are generated and stored in a linked list which makes reasoning much easier and generally ensures the code is readable. This does mean there is likely to be a small performance hit from this code, however, I haven't tested this and compared to the cost of sending the documents to Elasticsearch, it is likely insignificant. A linked list is used because it is a much more natural way of presenting the versions of the record than a normal list and it just worked when I conceptualised it that way instead of with several lists working in tandem with different index variables that were incremented and decremented accordingly.
Tests have been added to check that the bug is fixed and all the other tests relating to index ops have been fixed to work with the new index op generation function (the primary difference being that it now generates ops in ascending version order, not descending).

Closes: #34